requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
torch>=2.2.0
git+https://github.com/openai/CLIP.git
Pillow>=9.5.0
requests>=2.28.2
numpy>=1.24.2
